CSP-2399 Relocated implicit conversion tests from ProvidersControllerTests.cs to ProviderCourseModelTests.cs
