+++ src/SFA.DAS.Roatp.Api.UnitTests/Models/ProviderCourseModelTests.cs
@@ -1,8 +1,10 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using FluentAssertions;
 using NUnit.Framework;
 using SFA.DAS.Roatp.Application.ProviderCourse.Queries.GetProviderCourse;
 using SFA.DAS.Roatp.Domain.Entities;
+using SFA.DAS.Roatp.Domain.Models;
 
 namespace SFA.DAS.Roatp.Api.UnitTests.Models
 {
@@ -13,7 +15,7 @@ namespace SFA.DAS.Roatp.Api.UnitTests.Models
         [TestCase(false)]
         public void ProviderCourseOperator_ReturnsProviderCourseModel(bool hasPortableFlexiJobOption)
         {
-            var course = new ProviderCourse() { LarsCode = "1", HasPortableFlexiJobOption = hasPortableFlexiJobOption, Standard = new Standard() };
+            var course = new Domain.Entities.ProviderCourse() { LarsCode = "1", HasPortableFlexiJobOption = hasPortableFlexiJobOption, Standard = new Standard() };
             var model = (ProviderCourseModel)course;
 
             Assert.That(model, Is.Not.Null);
@@ -24,7 +26,7 @@ namespace SFA.DAS.Roatp.Api.UnitTests.Models
         [Test]
         public void ProviderCourseOperator_UpdateCourseInjectsExpectedValues()
         {
-            var course = new ProviderCourse() { LarsCode = "1", Standard = new Standard() };
+            var course = new Domain.Entities.ProviderCourse() { LarsCode = "1", Standard = new Standard() };
             var model = (ProviderCourseModel)course;
 
             var standardLookup = new Standard
@@ -56,11 +58,82 @@ namespace SFA.DAS.Roatp.Api.UnitTests.Models
                 locations.Add(new ProviderCourseLocation());
             }
 
-            var course = new ProviderCourse { Locations = locations, Standard = new Standard() };
+            var course = new Domain.Entities.ProviderCourse { Locations = locations, Standard = new Standard() };
             var model = (ProviderCourseModel)course;
 
             Assert.That(model, Is.Not.Null);
             Assert.AreEqual(expected, model.HasLocations);
+        }
+
+        [TestCase(true, "Apprenticeship")]
+        [TestCase(false, null)]
+        [TestCase(false, "NotARealCourseType")]
+        public void ProviderCourseModel_ImplicitConversion_FromDomainEntity_SetsProperties(bool hasLocations, string courseTypeString)
+        {
+            var providerCourse = new Domain.Entities.ProviderCourse
+            {
+                Id = hasLocations ? 10 : 20,
+                LarsCode = hasLocations ? "555" : "999",
+                StandardInfoUrl = hasLocations ? "http://standard" : null,
+                ContactUsEmail = hasLocations ? "a@b.c" : null,
+                ContactUsPhoneNumber = hasLocations ? "000" : null,
+                IsApprovedByRegulator = hasLocations ? true : (bool?)null,
+                IsImported = hasLocations,
+                HasPortableFlexiJobOption = hasLocations,
+                Locations = hasLocations ? [new Domain.Entities.ProviderCourseLocation()] : [],
+                Standard = courseTypeString == null ? null : new Standard
+                {
+                    CourseType = courseTypeString,
+                    IsRegulatedForProvider = string.Equals(courseTypeString, "Apprenticeship")
+                }
+            };
+
+            var model = (ProviderCourseModel)providerCourse;
+
+            model.Should().NotBeNull();
+            model.ProviderCourseId.Should().Be(providerCourse.Id);
+
+            // Locations
+            model.HasLocations.Should().Be(providerCourse.Locations.Count > 0);
+
+            // Standard / CourseType parsing
+            if (providerCourse.Standard == null)
+            {
+                model.IsRegulatedForProvider.Should().BeFalse();
+                model.CourseType.Should().BeNull();
+            }
+            else if (Enum.TryParse<CourseType>(providerCourse.Standard.CourseType, out var parsed))
+            {
+                model.CourseType.Should().Be(parsed);
+                model.IsRegulatedForProvider.Should().Be(providerCourse.Standard.IsRegulatedForProvider);
+            }
+            else
+            {
+                model.CourseType.Should().BeNull();
+                model.IsRegulatedForProvider.Should().Be(providerCourse.Standard.IsRegulatedForProvider);
+            }
+
+            // If present, other properties should map through
+            if (providerCourse.StandardInfoUrl != null)
+                model.StandardInfoUrl.Should().Be(providerCourse.StandardInfoUrl);
+            if (providerCourse.ContactUsEmail != null)
+                model.ContactUsEmail.Should().Be(providerCourse.ContactUsEmail);
+            if (providerCourse.ContactUsPhoneNumber != null)
+                model.ContactUsPhoneNumber.Should().Be(providerCourse.ContactUsPhoneNumber);
+            if (providerCourse.IsApprovedByRegulator.HasValue)
+                model.IsApprovedByRegulator.Should().Be(providerCourse.IsApprovedByRegulator);
+            model.IsImported.Should().Be(providerCourse.IsImported);
+            model.HasPortableFlexiJobOption.Should().Be(providerCourse.HasPortableFlexiJobOption);
+        }
+
+        [Test]
+        public void ProviderCourseModel_ImplicitConversion_WhenEntityIsNull(
+            )
+        {
+            Domain.Entities.ProviderCourse providerCourse = null;
+
+            var model = (ProviderCourseModel)providerCourse;
+            model.Should().BeNull();
         }
     }
 }
